# Simplify Repository.LoadBlob usage