--- a/internal/restic/repository.go
+++ b/internal/restic/repository.go
@@ -45,7 +45,7 @@ type Repository interface {
 	// new buffer will be allocated and returned.
 	LoadAndDecrypt(ctx context.Context, buf []byte, t FileType, id ID) (data []byte, err error)
 
-	LoadBlob(context.Context, BlobType, ID, []byte) (int, error)
+	LoadBlob(context.Context, BlobType, ID, []byte) ([]byte, error)
 	SaveBlob(context.Context, BlobType, []byte, ID) (ID, error)
 
 	LoadTree(context.Context, ID) (*Tree, error)