What is the purpose of this change? What does it change?
--------------------------------------------------------

Repository.LoadBlob currently requires its caller to compute the size of the buffer it needs. It then checks whether the buffer has the right size, and its helper loadBlob checks again. This leads to much duplicated, yet subtly different, code for computing sizes as well as repeated index lookups.

This PR changes LoadBlob to just return a freshly allocated buffer if insufficient space was passed in. The benchmarks show that fewer allocations have to be performed. The simplified code also makes it easier to add compression, since LoadBlob callers no longer need to know how big their buffers need to be (they'd have to know about compression as well as encryption).

```
name              old time/op    new time/op    delta
LoadTree-8           479µs ± 2%     481µs ± 2%    ~     (p=0.684 n=10+10)
LoadBlob-8          11.6ms ± 2%    11.6ms ± 2%    ~     (p=0.393 n=10+10)
LoadAndDecrypt-8    13.2ms ± 2%    13.3ms ± 3%    ~     (p=0.604 n=10+9)
LoadIndex-8         33.4ms ± 2%    33.2ms ± 1%    ~     (p=0.161 n=9+9)

name              old alloc/op   new alloc/op   delta
LoadTree-8          41.2kB ± 0%    41.1kB ± 0%  -0.23%  (p=0.000 n=10+10)
LoadBlob-8          2.28kB ± 0%    2.18kB ± 0%  -4.21%  (p=0.000 n=10+10)
LoadAndDecrypt-8    2.10MB ± 0%    2.10MB ± 0%    ~     (all equal)
LoadIndex-8         5.22MB ± 0%    5.22MB ± 0%    ~     (p=0.853 n=10+10)

name              old allocs/op  new allocs/op  delta
LoadTree-8             652 ± 0%       651 ± 0%  -0.15%  (p=0.000 n=10+10)
LoadBlob-8            24.0 ± 0%      23.0 ± 0%  -4.17%  (p=0.000 n=10+10)
LoadAndDecrypt-8      30.0 ± 0%      30.0 ± 0%    ~     (all equal)
LoadIndex-8          30.2k ± 0%     30.2k ± 0%    ~     (p=0.664 n=10+10)

name              old speed      new speed      delta
LoadBlob-8        86.2MB/s ± 2%  85.9MB/s ± 2%    ~     (p=0.393 n=10+10)
LoadAndDecrypt-8  75.7MB/s ± 2%  75.4MB/s ± 3%    ~     (p=0.603 n=10+9)
```

There is one functional change, which is the error message from LoadTree will now say "id %v not found in repository" rather than "tree %v not found in repository".

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------

Closes #2629.

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
- [x] I have added tests for all changes in this PR
- [x] I have run `gofmt` on the code in all commits
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [x] I'm done, this Pull Request is ready for review