@greatroar 2020-03-10
I also tested this with restic cat on a 4GB file. It's neither slower, nor faster. The SHA256 checks out.