@MichaelEischer 2020-03-10
So I guess this obsoletes #2629 ?