@greatroar 2020-04-22
I've merged with the earliest commit on master that resolved the merge conflict to get the CI working again. I hope that's ok.