@MichaelEischer 2020-04-22
Could you rebase the branch on the current master, that would lead to a much cleaner commit history (I could also do the rebase, if that's easier for you)? Sorry for being unclear on what I had in mind to resolve the merge conflict.