@MichaelEischer 2020-04-23
Merged, thanks! Github seems to have detected the rebase. Btw, your master branch seems to be slightly out of date.